requirements.txt
# ============================================================================
# FILE 6: requirements.txt
# LOCATION: requirements.txt (Project Root)
# PURPOSE: Python dependencies
# ============================================================================

# Core Framework
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# Deep Learning
tensorflow==2.15.0
keras==2.15.0

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Data Processing
numpy==1.24.3
pandas==2.1.3

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
requests==2.31.0

# Production Server
gunicorn==21.2.0

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0